WIP: рефакторинг организации компонентов
