+++ src/view/movie-card/movie-card.tpl.js
@@ -1,8 +1,7 @@
 import dayjs from 'dayjs';
-import Component from '../abstract/component';
-import { getFormattedRuntime, getSliceText } from '../utils/common';
+import { getFormattedRuntime, getSliceText } from '../../utils/common';
 
-const createMovieCard = (movie) => {
+export const createMovieCard = (movie) => {
   const { title, totalRating, release, runtime, genre, poster, description } = movie.filmInfo;
   const { alreadyWatched, favorite, watchlist } = movie.userDetails;
 
@@ -40,26 +39,3 @@ const createMovieCard = (movie) => {
   </article>
 `.trim();
 };
-
-export default class MovieCard extends Component {
-  #movie = null;
-
-  #linkClickHandler = (event) => {
-    event.preventDefault();
-    this._callback.click();
-  }
-
-  constructor(movie) {
-    super(movie);
-    this.#movie = movie;
-  }
-
-  get template() {
-    return createMovieCard(this.#movie);
-  }
-
-  setLinkClickHandler = (callback) => {
-    this._callback.click = callback;
-    this.element.querySelector('.film-card__link').addEventListener('click', this.#linkClickHandler);
-  }
-}
