+++ src/view/movie-counter/movie-counter.tpl.js
@@ -1,0 +1,3 @@
+export const createMovieCounter = (counter) => `
+  <p>${counter} movies inside</p>
+`.trim();
