+++ src/view/movies-page/movies-page.js
@@ -1,0 +1,9 @@
+import Component from '../../abstract/component';
+import { createMoviesPage } from './movies-page.tpl';
+
+export default class MoviesPage extends Component {
+
+  get template() {
+    return createMoviesPage();
+  }
+}
