+++ src/view/movies-page/movies-page.tpl.js
@@ -1,0 +1,1 @@
+export const createMoviesPage = () => '<section class="films"></section>';
