--- src/view/movies.js
@@ -1,10 +1,0 @@
-import Component from '../abstract/component';
-
-const createMoviesPage = () => '<section class="films"></section>';
-
-export default class MoviePage extends Component {
-
-  get template() {
-    return createMoviesPage();
-  }
-}
