+++ src/view/show-more-button/show-more-button.tpl.js
@@ -1,0 +1,1 @@
+export const createShowMore = () => '<button class="films-list__show-more">Show more</button>';
