+++ src/view/sorter/sorter.js
@@ -1,0 +1,8 @@
+import Component from '../../abstract/component';
+import { createSorter } from './sorter.tpl';
+
+export default class Sorter extends Component {
+  get template() {
+    return createSorter();
+  }
+}
